updates & use new endpoint
